+++ src/main/java/hu/smiths/dvdcomposer/model/FolderUtils.java
@@ -6,9 +6,9 @@ import org.apache.commons.io.FileUtils;
 
 public class FolderUtils {
 
-	public static Integer getFolderSize(File folder) {
+	public static Long getFolderSize(File folder) {
 		if (!folder.isDirectory())
 			throw new IllegalArgumentException(folder.getAbsolutePath() + " is not a folder!");
-		return (int) FileUtils.sizeOfDirectory(folder);
+		return FileUtils.sizeOfDirectory(folder);
 	}
 }
